requirements.txt
pytest
pyyaml
ansible
